requirements.txt
django-tinymce
easy_thumbnails